make viewer simpler
