+++ napari/_qt/__init__.py
@@ -1,5 +1,6 @@
 from ..resources import import_resources
 from .event_loop import gui_qt
+from .qt_main_window import Window
 from .widgets.qt_range_slider import QHRangeSlider, QVRangeSlider
 
 import_resources()
